第一个版本，简单完成了几个功能

+++ src/main/java/com/tianlin/dao/interfaces/ContentMapper.java
@@ -5,13 +5,13 @@ import java.util.List;
 import com.tianlin.entity.Content;
 
 public interface ContentMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Content record);
 
     int insertSelective(Content record);
 
-    Content selectByPrimaryKey(Integer id);
+    Content selectByPrimaryKey(Long id);
     
     /**多条件查询
      * **/
